xtf: Make the common hypercall header arch independent
Each architecture needs to define its own hypercall
handling interface. The hypercall handler should
be named as follows:
hypercall_<hypercall_name>
e.g. hypercall_console_io

The reason for that is to have a standard way
of calling hypercalls in the common code.

Modify the common <xtf/hypercall.h> header to
be architecture independent and move x86 handlers
to architecture specific code.
